fix solana sign#3
