+++ chain/solana/solana_client.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/gagliardetto/solana-go"
+	"github.com/gagliardetto/solana-go/rpc"
 	"log"
 	"sort"
 	"strings"
@@ -1069,4 +1071,24 @@ func (s *solclient) GetLatestBlockhash(commitmentType CommitmentType) (string, e
 	}
 
 	return blockhash, nil
+}
+
+// GetAccountInfo retrieves account information for a given token account
+func GetAccountInfo(sdkClient *rpc.Client, tokenAccount solana.PublicKey) (*rpc.GetAccountInfoResult, error) {
+	accountInfo, err := sdkClient.GetAccountInfo(context.Background(), tokenAccount)
+	if err != nil {
+		log.Println("Failed to get account info", "err", err)
+		return nil, err
+	}
+	return accountInfo, nil
+}
+
+// GetTokenSupply retrieves the token supply for a given mint public key
+func GetTokenSupply(sdkClient *rpc.Client, mintPubkey solana.PublicKey) (*rpc.GetTokenSupplyResult, error) {
+	tokenInfo, err := sdkClient.GetTokenSupply(context.Background(), mintPubkey, rpc.CommitmentFinalized)
+	if err != nil {
+		log.Println("Failed to get token supply", "err", err)
+		return nil, err
+	}
+	return tokenInfo, nil
 }
